# openapi/agentnest.openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "AgentNest Control Plane API",
    "version": "0.2.0",
    "description": "Lean tenant and business scoped AgentNest demo API."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18080"
    }
  ],
  "paths": {
    "/api/tasks": {
      "post": {
        "operationId": "submitTask",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAcceptedResponse"
                }
              }
            }
          },
          "202": {
            "description": "Task accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAcceptedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Scope denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Lifecycle state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "OpenClaw dispatch failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Dependency or model provider unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{taskId}": {
      "get": {
        "operationId": "getTask",
        "parameters": [
          {
            "name": "request_id",
            "in": "query",
            "required": true,
            "schema": {
              "minLength": 1,
              "maxLength": 128,
              "type": "string"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "minLength": 1,
              "maxLength": 128,
              "type": "string"
            }
          },
          {
            "name": "biz_domain",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^[A-Z][A-Z0-9_]*$",
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "minLength": 1,
              "maxLength": 128,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Scope denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Lifecycle state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "OpenClaw dispatch failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Dependency or model provider unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents": {
      "get": {
        "operationId": "listAgents",
        "parameters": [
          {
            "name": "request_id",
            "in": "query",
            "required": true,
            "schema": {
              "minLength": 1,
              "maxLength": 128,
              "type": "string"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "minLength": 1,
              "maxLength": 128,
              "type": "string"
            }
          },
          {
            "name": "biz_domain",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^[A-Z][A-Z0-9_]*$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Scope denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Lifecycle state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "OpenClaw dispatch failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Dependency or model provider unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/agents/{logicalAgentId}": {
      "get": {
        "operationId": "getAgent",
        "parameters": [
          {
            "name": "request_id",
            "in": "query",
            "required": true,
            "schema": {
              "minLength": 1,
              "maxLength": 128,
              "type": "string"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "minLength": 1,
              "maxLength": 128,
              "type": "string"
            }
          },
          {
            "name": "biz_domain",
            "in": "query",
            "required": true,
            "schema": {
              "pattern": "^[A-Z][A-Z0-9_]*$",
              "type": "string"
            }
          },
          {
            "name": "logicalAgentId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^tb_[a-f0-9]{20}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Scope denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Scoped resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Lifecycle state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal service error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "OpenClaw dispatch failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Dependency or model provider unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StandardErrorResponse": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "success",
          "code",
          "message",
          "request_id",
          "trace_id",
          "data",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "minLength": 1,
            "maxLength": 64,
            "type": "string"
          },
          "message": {
            "maxLength": 512,
            "type": "string"
          },
          "request_id": {
            "minLength": 1,
            "maxLength": 128,
            "type": "string"
          },
          "trace_id": {
            "minLength": 1,
            "maxLength": 128,
            "type": "string"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "minProperties": 1,
                "type": "object",
                "patternProperties": {
                  "^(.*)$": {}
                }
              },
              {
                "type": "null"
              }
            ]
          }
        }
      }
    }
  }
}